world_cup/requirements.txt
django==4.1.2
psycopg2-binary==2.9.4
djangorestframework==3.14.0
gunicorn==20.1.0
python-dotenv==0.19.2